# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service Booking API",
        "description": "Production-Quality Service Booking RESTful API in Laravel 12. This API serves a multi-user platform for managing service provider availability, enabling real-time booking by customers, and providing administrators with powerful reporting tools.",
        "contact": {
            "email": "admin@booking.api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/",
            "description": "Current Domain"
        }
    ],
    "paths": {
        "/api/v1/admin/reports/bookings-summary": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get bookings summary report",
                "description": "Get total bookings, cancellation rates, etc. with optional filtering",
                "operationId": "1916636a9f860c7eaaa838f79c6197d1",
                "parameters": [
                    {
                        "name": "provider_id",
                        "in": "query",
                        "description": "Filter by provider ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "01hw555666777"
                        }
                    },
                    {
                        "name": "service_id",
                        "in": "query",
                        "description": "Filter by service ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "01hw123456789"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter from date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter to date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Summary report retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "confirmed": {
                                                    "type": "integer",
                                                    "example": 120
                                                },
                                                "cancelled": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "completed": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "cancellation_rate": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0.133
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not an admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}